Implement a script to Clone VMs.
ESXi doesn't implement a high-level tool for cloning VMs, instead
assuming you'd work with a disk image directly. This tool, `vmkfstools`,
can work with both the base image or a snapshot if you use the source
argument correctly.

`clonevm.py` wraps this to provide a way to refer to _any_ disk image
that exists inside `/vmfs/volumes` based on the VM name and allows you
to refer to snapshots by number.

https://www.vmware.com/support/developer/vcli/vcli41/doc/reference/vmkfstools.html
https://recommender.vmware.com/solution/SOL-1623